+++ host/esutils/client_v6.go
@@ -18,7 +18,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package esutil
+package esutilss
 
 import (
 	"fmt"
